transcribe/requirements.txt
ffmpeg-python
openai-whisper
pydub
pyannote-audio
wave
numpy
scikit-learn
matplotlib